test(tracing): shared memory_exporter fixture + suite verification
Move the in-memory OTel exporter fixture to the root conftest.py so all tracing
tests share it. Full tracing set (16 tests) passes with no skips; representative
regression slice (1865 tests) green.
