+++ conftest.py
@@ -28,3 +28,30 @@ except (ImportError, OSError):
     ]:
         sys.modules.setdefault(_mod, MagicMock())
     sys.modules["mlx.core"] = _mock_mx
+
+
+import pytest
+
+
+@pytest.fixture
+def memory_exporter():
+    """Enable olmlx tracing against an in-memory OTel exporter; tear down after.
+
+    Shared by the tracing integration tests. Requires the [otel] dev deps
+    (skips cleanly otherwise).
+    """
+    pytest.importorskip("opentelemetry")
+    from opentelemetry.sdk.trace import TracerProvider
+    from opentelemetry.sdk.trace.export import SimpleSpanProcessor
+    from opentelemetry.sdk.trace.export.in_memory_span_exporter import (
+        InMemorySpanExporter,
+    )
+
+    import olmlx.utils.tracing as tracing
+
+    exporter = InMemorySpanExporter()
+    provider = TracerProvider()
+    provider.add_span_processor(SimpleSpanProcessor(exporter))
+    tracing.install_test_provider(provider)
+    yield exporter
+    tracing.shutdown_tracing()
